requirements.txt
flask==2.0.2
google-cloud-storage==1.42.3
